fix: address review — test-only preview builder, depth-exceeded redirects, isinstance dispatch
- Move matching_line_preview_of (the domain-Text builder, exercised only by
  tests) out of fragment_query_preview.py into
  fragment_query_preview_test_helpers.py. Production keeps only
  matching_line_preview_of_data.
- bibliography_documents_of now omits an id from bibliographyDocuments when
  resolving its redirect chain consumes the full MAX_REDIRECT_DEPTH hops and
  the final document is still deprecated, instead of returning a half-resolved
  record. A confirmed dangling target (query attempted, target doesn't exist)
  still resolves to the stored deprecated record as before.
- FragmentsQueryResource.on_get dispatches on
  isinstance(result, FragmentQueryResult) instead of casting based on the
  "limit" in query proxy.
